+++ src/ledger_editor/widgets/__init__.py
@@ -1,8 +1,7 @@
 """Textual widget components for the ledger editor."""
 
-from ledger_editor.widgets.balance_sidebar import BalanceSidebar
 from ledger_editor.widgets.filter_popup import FilterPopup
-from ledger_editor.widgets.register_panel import RegisterPanel
 from ledger_editor.widgets.transaction_table import JournalEditor
+from ledger_editor.widgets.view_filter_bar import ViewFilterBar
 
-__all__ = ["BalanceSidebar", "FilterPopup", "JournalEditor", "RegisterPanel"]
+__all__ = ["FilterPopup", "JournalEditor", "ViewFilterBar"]